Images for the First Version
This was a simple design using an Arduino Uno R3. We had most of the data displayed from the Serial Monitor from the Arduino IDE, but some things, such as temperature, showed up on an LCD panel. The main issue was that the Arduino board was not wifi capable, and there was a need for an onboard JST capable plug.